refactor - improved folder structure
