+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <meta-data
+        android:name="com.google.android.geo.API_KEY"
+        android:value="AIzaSyCM47WQ7fNA-y28PrxNOGtIk3xNv4Db6WY" />
+
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -13,7 +17,12 @@
         android:theme="@style/Theme.PickUpandDrop"
         tools:targetApi="31">
         <activity
-            android:name=".DeliveryDetails"
+            android:name=".delivery.presentation.locationAndMaps.ChooseDropLocationActivity"
+            android:exported="false"
+            android:label="@string/title_activity_choose_drop_location"
+            android:theme="@style/Theme.PickUpandDrop" />
+        <activity
+            android:name=".delivery.presentation.deliveryDetails.DeliveryDetailsActivity"
             android:exported="false"
             android:label="@string/title_activity_delivery_details"
             android:theme="@style/Theme.PickUpandDrop" />
