+++ app/src/main/java/com/example/pickupanddrop/common/Card.kt
@@ -116,165 +116,6 @@ fun ClickableCard(
     }
 }
 
-@Composable
-fun PickUpLocationCard(
-    modifier: Modifier = Modifier,
-    franchiseName: String,
-    franchiseLocation: String,
-    ownerName: String,
-    ownerPhoneNo: String
-) {
-
-    val labelFontWeight = MaterialTheme.typography.labelMedium.fontWeight
-
-    ElevatedCard(
-        modifier = modifier.border(1.dp, Color.Gray.copy(0.5f), RoundedCornerShape(8.dp)),
-        shape = RoundedCornerShape(8.dp),
-
-        ) {
-        Column() {
-            // pick up location and owner details
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(Color.Gray.copy(0.1f))
-                    .padding(8.dp)
-            ) {
-                Row {
-                    Text(
-                        text = "PICKUP LOCATION",
-                        fontSize = MaterialTheme.typography.labelMedium.fontSize,
-                        fontStyle = MaterialTheme.typography.labelMedium.fontStyle,
-                        fontWeight = labelFontWeight,
-                        color = Color.Gray
-                    )
-                    Text(text = " *", color = Color.Red.copy(0.8f))
-                }
-
-                Spacer(modifier = Modifier.height(12.dp))
-
-                Row {
-                    Text(
-                        text = "$franchiseName, ",
-                        fontWeight = MaterialTheme.typography.labelMedium.fontWeight
-                    )
-                    Text(
-                        text = franchiseLocation,
-                        fontStyle = MaterialTheme.typography.labelMedium.fontStyle,
-                        fontWeight = MaterialTheme.typography.labelMedium.fontWeight,
-                        color = Color.Gray,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                // owner name, phone no, and edit details option
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Icon(
-                        modifier = Modifier.size(16.dp),
-                        imageVector = Icons.Filled.Call,
-                        contentDescription = "phone no of owner",
-                        tint = Color.Gray
-                    )
-                    Spacer(modifier = Modifier.width(4.dp))
-                    // owner name and phone no
-                    Text(
-                        text = "$ownerName ($ownerPhoneNo)",
-                        color = Color.Gray,
-                        fontWeight = labelFontWeight,
-                        fontStyle = MaterialTheme.typography.labelMedium.fontStyle,
-                        fontSize = MaterialTheme.typography.labelMedium.fontSize
-                    )
-                    Spacer(modifier = Modifier.weight(1f))
-                    // edit details
-                    Text(
-                        modifier = Modifier.clickable { },
-                        text = "Edit Details",
-                        fontWeight = FontWeight(800),
-                        fontSize = MaterialTheme.typography.labelMedium.fontSize,
-                        color = Color.Blue.copy(0.6f)
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-
-            }
-
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(1.dp)
-                    .background(Color.Gray.copy(0.5f))
-            ) {}
-
-            val checkedStateForOTP = rememberSaveable {
-                mutableStateOf(false)
-            }
-            Row(
-                modifier = Modifier.padding(horizontal = 14.dp, vertical = 12.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Checkbox(
-                    modifier = Modifier.size(16.dp),
-                    checked = checkedStateForOTP.value, onCheckedChange = {
-                        checkedStateForOTP.value = it
-                    })
-                Spacer(modifier = Modifier.width(12.dp))
-                Text(
-                    text = "OTP for pickup",
-                    fontSize = MaterialTheme.typography.labelMedium.fontSize,
-                    fontWeight = labelFontWeight,
-                    fontStyle = MaterialTheme.typography.labelMedium.fontStyle
-                )
-            }
-        }
-
-    }
-}
-
-@Composable
-fun DropLocationCard(modifier: Modifier = Modifier) {
-    val blueBorderColor = Color.Blue.copy(0.7f)
-    ElevatedCard(
-        modifier = modifier.border(1.dp, blueBorderColor, RoundedCornerShape(8.dp))
-    ) {
-        Row(
-            modifier = Modifier.padding(14.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Column(modifier = Modifier.border(1.dp, blueBorderColor, CircleShape)) {
-                Icon(
-                    imageVector = Icons.Filled.Add,
-                    contentDescription = "add drop location",
-                    tint = blueBorderColor
-                )
-            }
-            Text(text = "Add drop Location")
-        }
-    }
-}
-
-@Preview
-@Composable
-fun DropLocationCardPreview() {
-    DropLocationCard()
-}
-
-@Preview
-@Composable
-fun PickUpLocationCardPreview() {
-    PickUpLocationCard(
-        franchiseName = "Petuk Ji G.Noida Franchise",
-        franchiseLocation = "IT waala digital success D, 296 Greater Noida Wandra",
-        ownerName = "Vinay",
-        ownerPhoneNo = "7970783256"
-    )
-}
-
 @Preview
 @Composable
 fun ClickableCardPreview() {
