+++ app/src/main/java/com/example/pickupanddrop/delivery/presentation/deliveryDetails/DeliveryDetailsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.pickupanddrop.presentation
+package com.example.pickupanddrop.delivery.presentation.deliveryDetails
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
@@ -13,8 +13,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.pickupanddrop.common.DropLocationCard
-import com.example.pickupanddrop.common.PickUpLocationCard
+import com.example.pickupanddrop.delivery.components.DropLocationCard
+import com.example.pickupanddrop.delivery.components.PickUpLocationCard
 
 @Composable
 fun DeliveryDetailsScreen() {
