+++ app/src/main/java/com/example/pickupanddrop/delivery/presentation/locationAndMaps/ChooseDropLocationActivity.kt
@@ -1,4 +1,4 @@
-package com.example.pickupanddrop
+package com.example.pickupanddrop.delivery.presentation.locationAndMaps
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,33 +10,21 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.pickupanddrop.presentation.DeliveryDetailsScreen
 import com.example.pickupanddrop.ui.theme.PickUpandDropTheme
 
-class DeliveryDetails : ComponentActivity() {
+class ChooseDropLocationActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             PickUpandDropTheme {
                 // A surface container using the 'background' color from the theme
-                DeliveryDetailsScreen()
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+
+                }
             }
         }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    PickUpandDropTheme {
-        Greeting("Android")
     }
 }
